# Use easyjson for JSON unmarshaling in createRegistryMessage