This commit improves the performance of JSON unmarshaling for each
message received by NATS. It adds a dependency on `easyjson`. We should
expect a ~4x improvement in performance when decoding NATS messages.

Before change:
```
go test -run Benchmark -bench=.
BenchmarkCreateRegistryMessage-4
100000             14882 ns/op PASS ok
code.cloudfoundry.org/gorouter/mbus     1.650s
```

After change:
```
go test -run Benchmark -bench=.
BenchmarkCreateRegistryMessage-4
300000              3908 ns/op PASS ok
code.cloudfoundry.org/gorouter/mbus     1.244s
```