@shalako 2017-05-23
Hello @flawedmatrix 
Could you provide a report showing performance improvement?

Thank you!